server-side/requirements.txt
aniso8601==8.1.0
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
ChatterBot @ git+https://github.com/Night9Knight/ChatterBot.git@28c80a6db13af12bac664dbd18c5d51ecc37210e
click==7.1.2
cymem==2.0.5
dnspython==2.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
Flask==1.1.2
Flask-RESTful==0.3.8
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mathparse==0.1.2
murmurhash==1.0.5
numpy==1.19.3
plac==1.1.3
preshed==3.0.5
pymongo==3.11.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
spacy==2.3.5
SQLAlchemy==1.3.22
srsly==1.0.5
thinc==7.4.5
tqdm==4.55.0
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.0
Werkzeug==1.0.1
zipp==3.4.0
